+++ progressive-db/src/main/java/de/tuda/progressive/db/statement/ProgressiveStatement.java
@@ -9,7 +9,7 @@ public interface ProgressiveStatement extends AutoCloseable {
 
 	ResultSetMetaData getMetaData();
 
-	long getReadPartitions();
+	int getReadPartitions();
 
 	boolean isDone();
