+++ progressive-db/src/main/java/de/tuda/progressive/db/statement/ProgressiveViewSelectStatement.java
@@ -72,7 +72,7 @@ public class ProgressiveViewSelectStatement implements ProgressiveStatement, Pro
 	}
 
 	@Override
-	public long getReadPartitions() {
+	public int getReadPartitions() {
 		return viewStatement.getReadPartitions();
 	}
